Extract root command and factory logic into separate packages
