+++ context/context.go
@@ -17,8 +17,6 @@ import (
 type Context interface {
 	Branch() (string, error)
 	SetBranch(string)
-	Remotes() (Remotes, error)
-	BaseRepo() (ghrepo.Interface, error)
 	SetBaseRepo(string)
 	Config() (config.Config, error)
 }
@@ -160,7 +158,6 @@ func New() Context {
 // A Context implementation that queries the filesystem
 type fsContext struct {
 	config   config.Config
-	remotes  Remotes
 	branch   string
 	baseRepo ghrepo.Interface
 }
@@ -194,60 +191,6 @@ func (c *fsContext) Branch() (string, error) {
 
 func (c *fsContext) SetBranch(b string) {
 	c.branch = b
-}
-
-func (c *fsContext) Remotes() (Remotes, error) {
-	if c.remotes == nil {
-		gitRemotes, err := git.Remotes()
-		if err != nil {
-			return nil, err
-		}
-		if len(gitRemotes) == 0 {
-			return nil, errors.New("no git remotes found")
-		}
-
-		sshTranslate := git.ParseSSHConfig().Translator()
-		resolvedRemotes := translateRemotes(gitRemotes, sshTranslate)
-
-		// determine hostname by looking at the "main" remote
-		var hostname string
-		if mainRemote, err := resolvedRemotes.FindByName("upstream", "github", "origin", "*"); err == nil {
-			hostname = mainRemote.RepoHost()
-		}
-
-		// filter the rest of the remotes to just that hostname
-		filteredRemotes := Remotes{}
-		for _, r := range resolvedRemotes {
-			if r.RepoHost() != hostname {
-				continue
-			}
-			filteredRemotes = append(filteredRemotes, r)
-		}
-		c.remotes = filteredRemotes
-	}
-
-	if len(c.remotes) == 0 {
-		return nil, errors.New("no git remote found for a github.com repository")
-	}
-	return c.remotes, nil
-}
-
-func (c *fsContext) BaseRepo() (ghrepo.Interface, error) {
-	if c.baseRepo != nil {
-		return c.baseRepo, nil
-	}
-
-	remotes, err := c.Remotes()
-	if err != nil {
-		return nil, err
-	}
-	rem, err := remotes.FindByName("upstream", "github", "origin", "*")
-	if err != nil {
-		return nil, err
-	}
-
-	c.baseRepo = rem
-	return c.baseRepo, nil
 }
 
 func (c *fsContext) SetBaseRepo(nwo string) {
